requirements.txt
anyio==3.6.2
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dadata==21.10.1
Django==4.1.3
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
PyJWT==2.6.0
pytz==2022.6
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.3
urllib3==1.26.13
